Read and write MyList from Firebase
Design: use "Map" in Firebase to represent the wish list since the only way to
update a list in Firebase is to replace the whole thing. But when we use
a Map we can just use the `database.getReference(KEY).push()` method to
add an item.

Testing: Run 'app', go to 'My List', add items, view updated list,
verified the same list is in Firebase