Fix My List margins
Design:
- default activities to not display an ActionBar by setting the theme to
NoActionBar. That way we can add our own Toolbar to each activity
individually
- don't display any Toolbar on MainActivity
- explicitly add the Toolbar so that we can constrain other elements in
the layout to it (i.e., make it so that the Toolbar doesn't cover up
things beneath it).
- add nested ContraintLayout to constrain the list view to the bottom of
the Toolbar
- add padding to list items instead of margin since margin creates space
between elements, while padding makes the actual element bigger. The
reason we don't want space in between elements is that that space is not
clickable. By using padding, the whole element (list item) fills up the
space (gets bigger), which means the user can tap anywhere on the item
to result in an "onClick" for that item.

Testing: tested on Moto g5 plus and Pixel 2 with 12 items in the list.
List becomes scrollable when there are too many to fit in the screen.